Refactor key entries structure
While key entries is provided as an object where the key corresponds to
the entrie's number, this complexity does not have to make it to the
template.

We already fill up empty entries by generating an array of key entries
length, hence no need to switch back to an object structure.