Stop overriding provison key for all devices
During formatting from device document to UI template, the `provision`
key was overriden regardless of the device's provisionability.

This triggered a UX bug when adding SUP devices with unknown
brand/model, as it prompted end-users to specify a MAC address
eventhough the device couldn't be provisioned.

Upon creating the device, kazoo would make a request to the provisioning
service, which would subsequently fail because of the missing
provisioning information.