Fix expression evaluation order
The original intent was to short-ciruit the evaluation of the second
operand when the first one was truthy.
However, since the second operand contains a ternary and no parenthesis
where used to group it, the first operand became part of the ternary
conditon, hence returning an incorect value when evaluating to true.